Code review follow up
